Add travisfile to build and push docker-images
Travis is a free-for-FOSS-projects CI pipeline service. Here, we use
travis to build and upload the docker-image of docker-matrix.

Furthermore, a simple test of the container is performed. This test
merely generates a config, waits for the container to start up, and
checks if it can connect to the version-endpoint of the API. More tests
can be added in this fashion in the future.

In this commit, there are 2 TODOs in the yaml-file. These need to be
filled by the maintainer using the travis cli client:
travis encrypt DOCKERHUB_USERNAME=<her/his user> --add env.global
travis encrypt DOCKERHUB_PASSWORD=<her/his pass> --add env.global

After that, the travis github-app needs to be activated for the
repository.